requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
numpy==1.24.1
pandas==1.5.2
pymongo==4.3.3
pyodbc==4.0.35
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
dnspython==2.2.1
